Update apiDeclaration.js to include exported interfaces and prettier formatting
In this commit, I have made changes to the apiDeclaration.js file to ensure that it now properly allows for the exporting of interfaces. This was a necessary update to ensure that our application's API functions as intended.

Additionally, I have added prettier formatting to the code, which helps to improve the overall readability and consistency of the file. This will make it easier for other developers to work with this code in the future.

Overall, this update should help to improve the functionality and maintainability of our application's API.